+++ src/components/footer/Footer.tsx
@@ -3,13 +3,12 @@ import Image from 'next/image';
 import Link from 'next/link';
 
 import logo from '@/assets/logo.jpg';
+import FooterLinks from '@/components/footer/FooterLinks';
 import NewsLetter from '@/components/footer/NewsLetter';
 import StatBlock from '@/components/footer/StatBlock';
 import constants from '@/constants';
 import footerLinks from '@/data/footer/footerLinks';
-import socialMedia from '@/data/footer/socialMedia';
 import statLinks from '@/data/footer/statLinks';
-import cn from '@/lib/cn';
 
 import type { Component } from '@/types';
 
@@ -35,27 +34,7 @@ const Footer: Component = () => (
             ))}
           </div>
 
-          <div className="flex space-x-4">
-            {socialMedia.map((social) => {
-              const { title, icon: Icon, url, color } = social;
-
-              return (
-                <a
-                  key={title}
-                  aria-label={title}
-                  href={url}
-                  rel="noopener noreferrer"
-                  target="_blank"
-                  className={cn(
-                    'rounded-lg bg-slate-800 p-3 text-gray-400 transition-all duration-300 hover:scale-110 hover:shadow-lg',
-                    color,
-                  )}
-                >
-                  <Icon className="size-5" />
-                </a>
-              );
-            })}
-          </div>
+          <FooterLinks />
         </div>
 
         <div className="grid grid-cols-1 gap-8 sm:grid-cols-2 lg:col-span-4 lg:grid-cols-4">
